fix: remove legacy atlas surfaces
