+++ layer/sessions/20260508-144836-859149000.md
@@ -1,0 +1,49 @@
+---
+type: session
+id: 20260508-144836-859149000
+runtime_id: 903f8d13-6fe5-4b29-b61c-53503fcf4ecc
+continuity_uid: 903f8d13-6fe5-4b29-b61c-53503fcf4ecc
+work_spec: unspecified
+title: pi session
+status: active
+llm: pi
+interface: pi
+created: 2026-05-08T18:48:36.859208+00:00
+updated: 2026-05-08T18:48:36.859208+00:00
+start_timestamp: 1778266116859
+source_log: /Users/nicabar/.pi/agent/sessions/--Users-nicabar-Projects-Sandbox-AI-RUST-patina--/2026-05-08T15-29-18-452Z_019e0835-1a74-735f-8ee9-cc668e98ec61.jsonl
+participants:
+- id: pi-51248
+  role: interface
+  interface: pi
+  interface_name: pi
+  display_name: nicabar
+  joined_at: 2026-05-08T18:48:36.859208+00:00
+interfaces:
+- pi
+git:
+  project_uid: 2bdc808e
+  branch: none
+  starting_commit: none
+  start_tag: session-20260508-144836-859149000-pi-start
+---
+
+## Previous Session Context
+<!-- AI: Summarize the last session from last-session.md -->
+
+## Goals
+- [ ] pi session
+
+## Activity Log
+### 14:48 - Session Start
+Session initialized with goal: pi session
+Working on branch: none
+Tagged as: session-20260508-144836-859149000-pi-start
+
+## Decisions
+
+## Evidence
+
+## Handoff
+
+## Outcome
