+++ mother/src/daemon_bootstrap_config.rs
@@ -184,17 +184,6 @@ mod tests {
                             &serde_json::json!({"version": "test"}),
                         )
                     }),
-                    get_atlas_dashboard: Arc::new(|_| crate::http_daemon::HttpResponse {
-                        status: 200,
-                        headers: vec![("Content-Type".to_string(), "text/html".to_string())],
-                        body: b"<html>atlas</html>".to_vec(),
-                    }),
-                    get_atlas_snapshot: Arc::new(|_| {
-                        crate::http_daemon::HttpResponse::json(
-                            200,
-                            &serde_json::json!({"summary": {"spec_count": 0}}),
-                        )
-                    }),
                     post_bridge_translate: Arc::new(|_| {
                         crate::http_daemon::HttpResponse::json(
                             200,
@@ -311,17 +300,6 @@ mod tests {
                         crate::http_daemon::HttpResponse::json(
                             200,
                             &serde_json::json!({"version": "test"}),
-                        )
-                    }),
-                    get_atlas_dashboard: Arc::new(|_| crate::http_daemon::HttpResponse {
-                        status: 200,
-                        headers: vec![("Content-Type".to_string(), "text/html".to_string())],
-                        body: b"<html>atlas</html>".to_vec(),
-                    }),
-                    get_atlas_snapshot: Arc::new(|_| {
-                        crate::http_daemon::HttpResponse::json(
-                            200,
-                            &serde_json::json!({"summary": {"spec_count": 0}}),
                         )
                     }),
                     post_bridge_translate: Arc::new(|_| {
