--- resources/pandos/atlas/pando.toml
@@ -1,11 +1,0 @@
-[pando]
-name = "atlas"
-description = "Mother-hosted visibility lens for specs, children, and toys"
-version = "0.1.0"
-
-[[children]]
-name = "legacy-typed-bridge"
-id = "legacy-typed-bridge"
-
-[composition]
-wiring = []
