--- ui/atlas/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "patina-atlas-ui",
-  "version": "0.0.1",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "dev": "vite dev",
-    "build": "vite build",
-    "preview": "vite preview"
-  },
-  "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^3.1.2",
-    "svelte": "^4.2.18",
-    "vite": "^5.4.8"
-  }
-}
