--- ui/atlas/src/main.ts
@@ -1,7 +1,0 @@
-import App from './App.svelte';
-
-const app = new App({
-  target: document.getElementById('app')!
-});
-
-export default app;
